my_env/pyproject.toml
[project]
name = "acde-openenv"
version = "1.0.0"
description = "Adaptive Crisis Decision Environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn==0.32.1",
  "pydantic==2.10.3",
  "requests==2.32.3",
  "openai>=2.7.2",
  "openenv-core"
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*", "server*"]
exclude = ["data*"]
